refactor: 점수 비교 메서드 수정
